# openapi.yaml
openapi: 3.0.0

info:
  title: Mood Joke Generator API
  version: 2.0.0
  description: |
    An agentic joke API powered by Claude. Submit a mood in plain natural language
    and an AI agent will interpret it, query the joke datastore, and return the
    most relevant joke. Rate jokes to build an aggregate feedback dataset.

    ## How the agent works
    `POST /get-joke` does not do a simple keyword lookup. The raw mood string is
    passed to a Claude agent that:
    1. Interprets the emotional context using natural language understanding
    2. Chooses its own short category label for the mood (no fixed list)
    3. Calls an internal `pick_jokes` tool to retrieve candidate jokes from the
       configured datastore (flat text file or vector database)
    4. Selects and returns the single most appropriate joke

    ## Datastore backends
    The active backend is controlled by the `STORE` environment variable on the
    server:
    - `local` — reads from a flat `randomJokes.txt` file, one joke per line (default)
    - `vector` — queries a vector database for semantic similarity search

  contact:
    name: Mood Joke Generator
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Jokes
    description: Retrieve AI-curated jokes based on natural language mood input
  - name: Ratings
    description: Submit and retrieve joke ratings
  - name: Stats
    description: Aggregate rating statistics
  - name: Usage
    description: Anthropic API token consumption and estimated cost for this server process

paths:

  # ─── POST /get-joke ────────────────────────────────────────────────────────
  /get-joke:
    post:
      operationId: getJoke
      tags: [Jokes]
      summary: Get a joke matched to a mood
      description: |
        Accepts a free-text mood description and returns the most contextually
        appropriate joke from the datastore, selected by an AI agent.

        The `mood` field accepts any natural language description — the agent
        invents a category label on the fly. Examples:
        - `"happy"` — agent picks a category like `happy`
        - `"feeling down after a rough day"` — agent picks something like `sad`
        - `"completely wiped out from jet lag"` — agent picks something like `tired`
        - `"my presentation just went sideways"` — agent picks something like `stressed`

        The agent always calls the internal `pick_jokes` tool before selecting a
        joke, so candidates are drawn from the real datastore rather than
        generated by the LLM.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetJokeRequest'
            examples:
              simple_mood:
                summary: Exact category keyword
                value:
                  mood: happy
              natural_language_tired:
                summary: Natural language — tired
                value:
                  mood: completely wiped out from jet lag
              natural_language_stressed:
                summary: Natural language — stressed
                value:
                  mood: my presentation just went sideways and I want to hide
              natural_language_sad:
                summary: Natural language — sad
                value:
                  mood: feeling really down after a rough week
      responses:
        '200':
          description: A joke was successfully selected by the agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetJokeResponse'
              examples:
                success:
                  summary: Successful joke response
                  value:
                    joke: Why don't scientists trust atoms? Because they make up everything!
                    mood: happy
                natural_language:
                  summary: Natural language mood mapped by the agent
                  value:
                    joke: I stayed up all night to see where the sun went. Then it dawned on me.
                    mood: completely wiped out from jet lag
        '400':
          description: The `mood` field was missing, empty, or whitespace-only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_mood:
                  summary: Missing mood field
                  value:
                    error: mood is required
                empty_mood:
                  summary: Empty string
                  value:
                    error: mood is required
        '500':
          description: The AI agent encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                agent_error:
                  summary: Agent failure
                  value:
                    error: The joke agent encountered an error. Please try again.

  # ─── POST /rate-joke ───────────────────────────────────────────────────────
  /rate-joke:
    post:
      operationId: rateJoke
      tags: [Ratings]
      summary: Submit a rating for a joke
      description: |
        Records a 1–3 star rating for a previously returned joke. The `mood` and
        `joke` fields are echoed back from the `/get-joke` response so each
        rating is linked to its context.

        Ratings are stored in memory on the server and can be retrieved via
        `GET /stats`. The rating scale is:
        - `1` — Meh (😐)
        - `2` — Good (🙂)
        - `3` — Hilarious (😂)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateJokeRequest'
            examples:
              hilarious:
                summary: Highest rating
                value:
                  mood: happy
                  joke: Why don't scientists trust atoms? Because they make up everything!
                  rating: 3
              meh:
                summary: Lowest rating
                value:
                  mood: tired
                  joke: I'm reading a book on anti-gravity. It's impossible to put down!
                  rating: 1
      responses:
        '200':
          description: Rating recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateJokeResponse'
              examples:
                success:
                  value:
                    message: Thanks for your feedback!

  # ─── GET /usage ────────────────────────────────────────────────────────────
  /usage:
    get:
      operationId: getUsage
      tags: [Usage]
      summary: Get Anthropic API usage and estimated cost
      description: |
        Returns cumulative Anthropic API usage for the current server process:
        token counters (input / output / cache create / cache read), estimated
        cost in USD based on configured per-MTok pricing, the latest rate-limit
        snapshot from Anthropic response headers, and session info.

        ## Reset behaviour
        Counters are in-memory only and reset when the server restarts. For
        authoritative historical usage, query the Anthropic Admin API:
        `/v1/organizations/usage_report/messages`.

        ## "Balance credit" note
        Anthropic does not expose a remaining-credit-balance endpoint via the
        public API. The `rateLimits` section is the closest equivalent — it
        reflects the request/token windows returned with the most recent
        response (including reset timestamps).

        ## Pricing
        Defaults to `claude-sonnet-4-6` list pricing as of 2026. Override via
        env vars: `PRICE_INPUT_PER_MTOK`, `PRICE_OUTPUT_PER_MTOK`,
        `PRICE_CACHE_WRITE_PER_MTOK`, `PRICE_CACHE_READ_PER_MTOK`.
      responses:
        '200':
          description: Usage snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
              examples:
                with_usage:
                  summary: After several joke requests
                  value:
                    totals:
                      requests: 6
                      input_tokens: 4820
                      output_tokens: 612
                      cache_creation_input_tokens: 0
                      cache_read_input_tokens: 0
                      total_tokens: 5432
                      estimated_cost_usd: 0.023640
                    byModel:
                      claude-sonnet-4-6:
                        requests: 6
                        input_tokens: 4820
                        output_tokens: 612
                        cache_creation_input_tokens: 0
                        cache_read_input_tokens: 0
                        total_tokens: 5432
                        estimated_cost_usd: 0.023640
                    pricing_per_mtok_usd:
                      input: 3.00
                      output: 15.00
                      cache_write: 3.75
                      cache_read: 0.30
                    rateLimits:
                      lastUpdatedAt: "2026-05-26T14:30:00.000Z"
                      requests:
                        limit: 50
                        remaining: 44
                        resetAt: "2026-05-26T14:31:00Z"
                      tokens:
                        limit: 50000
                        remaining: 44568
                        resetAt: "2026-05-26T14:31:00Z"
                      input_tokens:
                        limit: 40000
                        remaining: 35180
                        resetAt: "2026-05-26T14:31:00Z"
                      output_tokens:
                        limit: 10000
                        remaining: 9388
                        resetAt: "2026-05-26T14:31:00Z"
                    session:
                      startedAt: "2026-05-26T14:00:00.000Z"
                      uptimeSeconds: 1800
                    notes:
                      - Token counts and costs are tracked per-process and reset on server restart.
                      - Costs are estimates calculated from the configured per-MTok pricing.
                      - Anthropic does not expose a remaining-credit-balance endpoint via the public API.
                      - For authoritative historical usage, query the Anthropic Admin API.
                fresh_server:
                  summary: Server just started — no requests yet
                  value:
                    totals:
                      requests: 0
                      input_tokens: 0
                      output_tokens: 0
                      cache_creation_input_tokens: 0
                      cache_read_input_tokens: 0
                      total_tokens: 0
                      estimated_cost_usd: 0
                    byModel: {}
                    pricing_per_mtok_usd:
                      input: 3.00
                      output: 15.00
                      cache_write: 3.75
                      cache_read: 0.30
                    rateLimits: null
                    session:
                      startedAt: "2026-05-26T14:00:00.000Z"
                      uptimeSeconds: 5

  # ─── GET /stats ────────────────────────────────────────────────────────────
  /stats:
    get:
      operationId: getStats
      tags: [Stats]
      summary: Get aggregate rating statistics
      description: |
        Returns the total number of ratings submitted, the average rating score
        across all jokes, and the full log of individual rating entries.

        `averageRating` is returned as a string formatted to 2 decimal places
        (e.g. `"2.33"`), or `null` when no ratings have been submitted yet.

        Note: ratings are stored in server memory and reset on server restart.
      responses:
        '200':
          description: Aggregate statistics and full rating log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
              examples:
                with_ratings:
                  summary: Stats with existing ratings
                  value:
                    totalRatings: 3
                    averageRating: "2.33"
                    ratings:
                      - mood: happy
                        joke: Why don't scientists trust atoms? Because they make up everything!
                        rating: 3
                        time: "2025-06-01T10:00:00.000Z"
                      - mood: tired
                        joke: I'm reading a book on anti-gravity. It's impossible to put down!
                        rating: 2
                        time: "2025-06-01T10:05:00.000Z"
                      - mood: sad
                        joke: Why don't skeletons fight each other? They don't have the guts.
                        rating: 2
                        time: "2025-06-01T10:10:00.000Z"
                no_ratings:
                  summary: No ratings submitted yet
                  value:
                    totalRatings: 0
                    averageRating: null
                    ratings: []

# ─── Components ───────────────────────────────────────────────────────────────
components:
  schemas:

    # ── Requests ──────────────────────────────────────────────────────────────

    GetJokeRequest:
      type: object
      required: [mood]
      properties:
        mood:
          type: string
          minLength: 1
          description: |
            A free-text description of the user's current emotional state.
            Accepts any natural language — the AI agent will interpret it and
            map it to the most relevant jokes in the datastore.
          examples:
            - happy
            - feeling really down after a tough week
            - completely wired and can't sit still
            - wiped out after a red-eye flight

    RateJokeRequest:
      type: object
      required: [mood, joke, rating]
      properties:
        mood:
          type: string
          minLength: 1
          description: The mood string originally sent to `POST /get-joke`
          example: happy
        joke:
          type: string
          minLength: 1
          description: The joke text returned by `POST /get-joke`
          example: Why don't scientists trust atoms? Because they make up everything!
        rating:
          oneOf:
            - type: integer
            - type: string
          description: |
            Rating value on a 1–3 scale. Accepts either an integer or a
            numeric string — the server coerces it to a number.
            - `1` — Meh
            - `2` — Good
            - `3` — Hilarious
          enum: [1, 2, 3, "1", "2", "3"]
          example: 3

    # ── Responses ─────────────────────────────────────────────────────────────

    GetJokeResponse:
      type: object
      required: [joke, mood]
      properties:
        joke:
          type: string
          description: The joke selected by the AI agent from the datastore
          example: Why don't scientists trust atoms? Because they make up everything!
        mood:
          type: string
          description: The raw mood string that was submitted in the request
          example: happy

    RateJokeResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Confirmation message
          example: Thanks for your feedback!

    StatsResponse:
      type: object
      required: [totalRatings, averageRating, ratings]
      properties:
        totalRatings:
          type: integer
          minimum: 0
          description: Total number of ratings recorded since the server started
          example: 42
        averageRating:
          type:
            - string
            - "null"
          description: |
            Mean rating across all entries, formatted to 2 decimal places.
            Returns `null` when no ratings have been submitted.
          example: "2.33"
        ratings:
          type: array
          description: Full log of individual rating entries
          items:
            $ref: '#/components/schemas/RatingEntry'

    RatingEntry:
      type: object
      required: [mood, joke, rating, time]
      properties:
        mood:
          type: string
          description: The mood submitted with the original joke request
          example: happy
        joke:
          type: string
          description: The joke text that was rated
          example: Why don't scientists trust atoms? Because they make up everything!
        rating:
          type: integer
          minimum: 1
          maximum: 3
          description: The submitted rating (1 = Meh, 2 = Good, 3 = Hilarious)
          example: 3
        time:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the rating was recorded
          example: "2025-06-01T10:00:00.000Z"

    # ── Usage ─────────────────────────────────────────────────────────────────

    UsageCounters:
      type: object
      required:
        [requests, input_tokens, output_tokens, cache_creation_input_tokens, cache_read_input_tokens]
      properties:
        requests:                    { type: integer, minimum: 0, description: "Total Anthropic API requests issued" }
        input_tokens:                { type: integer, minimum: 0, description: "Non-cached input tokens" }
        output_tokens:               { type: integer, minimum: 0, description: "Output tokens generated by the model" }
        cache_creation_input_tokens: { type: integer, minimum: 0, description: "Tokens written to the prompt cache" }
        cache_read_input_tokens:     { type: integer, minimum: 0, description: "Tokens read from the prompt cache" }
        total_tokens:                { type: integer, minimum: 0, description: "Sum of all token categories above" }
        estimated_cost_usd:
          type: number
          minimum: 0
          description: "Cost in USD, computed from configured per-MTok pricing"

    RateLimitWindow:
      type: object
      description: "A single Anthropic rate-limit window — request, token, input-token, or output-token."
      properties:
        limit:
          type:
            - integer
            - "null"
          description: "Maximum allowed in this window"
        remaining:
          type:
            - integer
            - "null"
          description: "Remaining quota in this window — the closest signal to a 'credit balance'"
        resetAt:
          type:
            - string
            - "null"
          format: date-time
          description: "RFC 3339 timestamp at which the window resets"

    UsageResponse:
      type: object
      required: [totals, byModel, pricing_per_mtok_usd, rateLimits, session, notes]
      properties:
        totals:
          $ref: '#/components/schemas/UsageCounters'
        byModel:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/UsageCounters'
          description: "Counters bucketed by the model returned by Anthropic"
        pricing_per_mtok_usd:
          type: object
          required: [input, output, cache_write, cache_read]
          properties:
            input:       { type: number, description: "USD per 1M non-cached input tokens" }
            output:      { type: number, description: "USD per 1M output tokens" }
            cache_write: { type: number, description: "USD per 1M cache-creation tokens (5-min TTL)" }
            cache_read:  { type: number, description: "USD per 1M cache-read tokens" }
        rateLimits:
          oneOf:
            - type: "null"
            - type: object
              properties:
                lastUpdatedAt:
                  type: string
                  format: date-time
                  description: "When the snapshot was captured"
                requests:      { $ref: '#/components/schemas/RateLimitWindow' }
                tokens:        { $ref: '#/components/schemas/RateLimitWindow' }
                input_tokens:  { $ref: '#/components/schemas/RateLimitWindow' }
                output_tokens: { $ref: '#/components/schemas/RateLimitWindow' }
                retry_after_seconds:
                  type: integer
                  description: "Present only when the last response was rate-limited"
          description: "Latest Anthropic rate-limit header snapshot. `null` until at least one API call has completed."
        session:
          type: object
          required: [startedAt, uptimeSeconds]
          properties:
            startedAt:     { type: string, format: date-time }
            uptimeSeconds: { type: integer, minimum: 0 }
        notes:
          type: array
          items: { type: string }
          description: "Caveats about what these numbers do and do not represent."

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable description of what went wrong
          example: mood is required